pyproject.toml
[tool.poetry]
name = "ui-agent"
version = "1.0.0"
description = "UI Agent"
authors = ["Scalling Intelligence <engineering@scaling.ai>"]
readme = "README.md"
packages = [
    { include = "common" },
    { include = "workflow" },
]
include = ["pyproject.toml", "poetry.lock"]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
pydantic = "2.11.4"
python-dotenv = "1.1.0"
toml = ">=0.10.2,<0.11.0"
colorlog = "^6.9.0"
arrow = "^1.3.0"

[tool.poetry.group.agent]
optional = true

[tool.poetry.group.agent.dependencies]
jinja2 = ">=3.1.6,<4.0.0"
litellm = ">=1.67.5,<2.0.0"
termcolor = ">=3.1.0,<4.0.0"
tenacity = ">=9.1.2,<10.0.0"
json-repair = ">=0.44.1,<0.45.0"
boto3 = ">=1.36.20,<1.36.24 || >=1.37.0,<1.37.4"
botocore = ">=1.36.20,<1.36.24 || >=1.37.0,<1.37.4"
pylcs = ">=0.1.1,<0.2.0"
tree-sitter-python = "^0.23.6"
setuptools = ">=80.1.0,<81.0.0"
python-json-logger = ">=3.3.0,<4.0.0"
uvicorn = ">=0.34.2,<0.35.0"
redis = "^6.0.0"
exa-py = "^1.13.1"
daytona-sdk = "^0.18.1"
langgraph = "^0.4.7"
langchain = { extras = ["anthropic"], version = "^0.3.25" }
psycopg = { extras = ["binary", "pool"], version = "^3.2.9" }
prompt-toolkit = "^3.0.51"
python-frontmatter = "^1.1.0"
openai = "^1.99.6"

[tool.poetry.group.modem]
optional = true

[tool.poetry.group.modem.dependencies]
celery = "5.5.2"
kombu = "5.5.3"
pyopenssl = "25.0.0"
flower = "2.0.1"
websocket-client = "1.8.0"

[tool.poetry.group.gateway]
optional = true

[tool.poetry.group.gateway.dependencies]
fastapi = "0.115.12"
httpx = { extras = ["socks"], version = "^0.28.1" }


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
black = "^25.1.0"
isort = "^6.0.1"
autoflake = "^2.3.1"


[tool.poetry.group.workflow.dependencies]
tweepy = "^4.16.0"

[tool.isort]
profile = "black"
line_length = 88
float_to_top = true

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
